% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayefdr-package.R
\docType{package}
\name{bayefdr-package}
\alias{bayefdr-package}
\alias{bayefdr}
\title{The 'nibbles' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Detecting differential gene expression with a semiparametric hierarchical 
mixture method
Michael A. Newton, Amine Noueiry, Deepayan Sarkar, Paul Ahlquist
https://doi.org/10.1093/biostatistics/5.2.155
}
