% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayefdr.R
\name{efdr}
\alias{efdr}
\alias{efnr}
\title{EFDR and EFNR estimation}
\usage{
efdr(evidence_threshold, probs)

efnr(evidence_threshold, probs)
}
\arguments{
\item{evidence_threshold}{Scalar value specifying the evidence threshold
at which the EFDR or EFNR should be evaluated.}

\item{probs}{Vector of probabilities.}
}
\description{
Calculate the Expected False Discovery Rate (EFDR) or Expected False 
Negative Rate (EFNR) in a vector of probabilities, given a specified
evidence threshold.
}
