% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayefdr.R
\name{optimal}
\alias{optimal}
\title{Retrieve the index of the optimal probability threshold.}
\usage{
optimal(x)
}
\arguments{
\item{x}{An object of class "bayefdr".}
}
\value{
The integer index of the optimal probability threshold.
}
\description{
Retrieve the index of the optimal probability threshold.
}
