% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayefdr.R
\name{plot.bayefdr}
\alias{plot.bayefdr}
\title{Plot the EFDR, EFNR grids of a bayefdr object.}
\usage{
\method{plot}{bayefdr}(x, ...)
}
\arguments{
\item{x}{An object of class bayefdr.}

\item{...}{Unused.}
}
\value{
A ggplot.
}
\description{
Plot the EFDR, EFNR grids of a bayefdr object.
}
