% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayefdr.R
\name{print.bayefdr}
\alias{print.bayefdr}
\alias{head.bayefdr}
\title{Print methods for bayefdr objects.}
\usage{
\method{print}{bayefdr}(x, ...)

\method{head}{bayefdr}(x, ...)
}
\arguments{
\item{x}{An object of class bayefdr.}

\item{...}{Unused.}
}
\description{
Print methods for bayefdr objects.
}
